backend/log_parser/requirements.txt
minio
psycopg2-binary
python-dotenv
